Moved function to own file and mocked it in tests.
