+++ src/server/config.ts
@@ -4,6 +4,7 @@ import https from 'https';
 import * as jose from 'jose';
 import { FeatureToggle } from '../universal/config/feature-toggles';
 import { IS_DEVELOPMENT, IS_OT, IS_TAP } from '../universal/config/env';
+import { getFromEnv } from './helpers/env';
 import { jsonCopy } from '../universal/helpers/utils';
 import { TokenData } from './helpers/app';
 import fs from 'fs';
@@ -34,24 +35,6 @@ export function getCert(envVarName: string) {
   return IS_DEVELOPMENT
     ? getCertificateSync(envVarName)
     : decodeBase64EncodedCertificateFromEnv(envVarName);
-}
-
-/** Retrieve an environment variable.
- *
- * - Will never return undefined when `isRequired` (defaults `true`).
- * - Throws an error when a variable doesn't exist and it `isRequired`.
- */
-export function getFromEnv(
-  key: string,
-  isRequired: boolean = true
-): string | undefined {
-  if (key in process.env) {
-    return process.env[key];
-  }
-  if (isRequired) {
-    throw new Error(`ENV undefined key: ${key}.`);
-  }
-  console.warn(`ENV undefined, but not required: ${key}`);
 }
 
 export const IS_DEBUG = process.env.DEBUG === '1';
